requirements.txt
mysql-connector-python==8.3.0
numpy==1.23.5
pandas==2.2.0
pillow==10.2.0
plotly==5.9.0
streamlit==1.31.1
streamlit-option-menu==0.3.12
yfinance==0.2.36